requirements.txt
Django
Pillow
psycopg2
Faker
django-ckeditor
django-filter